Fixed a compiler warning
git-svn-id: https://svn.aps.anl.gov/AOP/oag/trunk/apps/src/elegant@16274 c1b61890-85f3-ea29-e4d6-a272fe79f168

+++ recycle.c
@@ -78,7 +78,7 @@ char   *purpose;
   if (!x)
   {
     fprintf(stderr, "malloc of %d failed for %s\n", 
-	    len, purpose);
+	    (int)len, purpose);
     exit(SUCCESS);
   }
   return x;
